Add active user component
Active user component was created. The component can be
reached by clicking on a profile icon. Feature includes
sliding animation when going to next active user. In this
snapshot the code for the team section still needs to be
cleaned and has yet to merge with up to date master.

Feature was added to display user details such as a
description and social networks.