Create animation class
Moved animation code into a class called ScrollAnim.
Class has a function to initialize TimelineLite
animations. The class also includes functions to play
slide down, right, up and left animations.

Purpose of this feature was to seperate code for animation
from UserList and UserImg. This makes it easier to debug or change
animations.