Tony: Changed App.scss and Team.scss
On line 136 - 166 the class .team-bg was added next
to #team. When designing the component I made the
background a seperate div from the content. The
background had position: absolute, so it needed the
height to be specified. This change fixes that. Also removed
height property from .team-bg inside Team.scss.